账本列表增加收入支出统计

+++ account-account-service/src/main/java/com/sxkl/webapp/account/accountbook/entity/Account.java
@@ -42,8 +42,8 @@ public class Account {
 	private String userId;
 
 	@Transient
-	private double income;
+	private float income;
 	
 	@Transient
-	private double outcome;
+	private float outcome;
 }
